chore: renamed mocks folder to __mocks__
