+++ packages/shared/lib/tests/__mocks__/platform.ts
