+++ packages/shared/lib/tests/currency.test.ts
@@ -1,4 +1,4 @@
-import './mocks/matchMedia'
+import './__mocks__/matchMedia'
 
 import { convertFromFiat, convertToFiat, isFiatCurrency } from '../currency'
 import { AvailableExchangeRates, CurrencyTypes } from '../typings/currency'
