+++ packages/shared/lib/tests/time.test.ts
@@ -1,5 +1,5 @@
-import './mocks/i18n'
-import './mocks/matchMedia'
+import './__mocks__/i18n'
+import './__mocks__/matchMedia'
 
 import {
     getBestTimeDuration,
