+++ packages/shared/lib/tests/utils.test.ts
@@ -1,4 +1,4 @@
-import './mocks/matchMedia'
+import './__mocks__/matchMedia'
 
 import { migrateObjects, range } from '../utils'
